Creating a Custom Authentication Hook
