+++ src/App.js
@@ -1,5 +1,5 @@
 /* eslint-disable react/jsx-pascal-case */
-import React, { useState, useCallback, useEffect } from "react";
+import React from "react";
 import {
   BrowserRouter as Router,
   Route,
@@ -15,62 +15,11 @@ import P_Places_PlaceId_Edit from "./pages/places/_place_id/edit";
 import P_Places_New from "./pages/places/new";
 import P_Error from "./pages/error";
 import { AuthProvider } from "./context/auth";
+import { useAuth } from "./hooks/auth";
 import "./assets/styles.css";
 
-let logoutTimer;
-
 const App = () => {
-  const [token, setToken] = useState(null);
-  const [tokenExpirationDate, setTokenExpirationDate] = useState();
-  const [userId, setUserId] = useState(false);
-
-  const login = useCallback((uid, token, expirationDate) => {
-    setToken(token);
-    setUserId(uid);
-    const tokenExpirationDate =
-      expirationDate || new Date(new Date().getTime() + 1000 * 60 * 60);
-    setTokenExpirationDate(tokenExpirationDate);
-    localStorage.setItem(
-      "userData",
-      JSON.stringify({
-        userId: uid,
-        token,
-        expirationDate: tokenExpirationDate.toISOString(),
-      }),
-    );
-  }, []);
-  const logout = useCallback(() => {
-    setToken(null);
-    setTokenExpirationDate(null);
-    setUserId(null);
-    localStorage.removeItem("userData");
-  }, []);
-
-  useEffect(() => {
-    if (token && tokenExpirationDate) {
-      const remainingTiming =
-        tokenExpirationDate.getTime() - new Date().getTime();
-      logoutTimer = setTimeout(logout, remainingTiming);
-    } else {
-      clearTimeout(logoutTimer);
-    }
-  }, [logout, token, tokenExpirationDate]);
-
-  useEffect(() => {
-    const storedData = JSON.parse(localStorage.getItem("userData"));
-    if (
-      storedData &&
-      storedData.token &&
-      new Date(storedData.expirationDate) > new Date()
-    ) {
-      login(
-        storedData.userId,
-        storedData.token,
-        new Date(storedData.expirationDate),
-      );
-    }
-  }, [login]);
-
+  const [token, login, logout, userId] = useAuth();
   return (
     <AuthProvider
       value={{ isLoggedIn: Boolean(token), token, userId, login, logout }}
